Reading list
This app takes user input and asks to choose what you would like to do: add, see, or quit. 
you can add a new book with a book name, author, and year of publication. It saves the new entry in a CSV file then we can choose to list all books.