Decimal to binary conversion
This program covert decimal numbers to binary and Hexa. The user is asked to input a value for a conversion and then numbers are converted to this entered number. Before the actual result, used also enter a range to check the conversion, if it is correct or not.
This program was built to learn lists, loops, and other basic python functionalities.