Word Guessing Game
This is a word guessing game. In this game, the User is asked to guess a word from a randomly chosen category like countries, classes, fruits, etc. If the user enters a wrong input, a hint is shown to the user in the form of a letter and this continues till the word is found!