Python oops Program
This is a complete program using oops concepts in python. It has three classes, Menu, Business, and Franchise. 
With Menu, we create different menus with different timing.
With Business class, Different Business and menus are created. 
And with Franchise class, different new franchises can be created. 
And then we can find their timings and menus.