Loan Calculator
This is a loan calculator program. The user is asked to enter the loan amount, interest rate along with monthly payment. And through a series of calculations, the loan amount is calculated and a final summary is presented to the user. The final summary also shows interest paid with loan amount and total number of months.